+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - CIP68 `getMetadataFromOutputDatum` parsing of `files` and custom fields
 
+### Removed
+
+- `addr_vk` is no longer valid address type as returned by `detectAndValidateAddressType`
+
 ## 2.3.0 - 2023-02-02
 
 ### Added