+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support for ScriptHash payment credential using `script` addresses.
 
+### Changed
+
+- do not leak framework in errors
+
 ## 2.3.1 - 2023-02-06
 
 ### Fixed